move image related class to image package
