+++ pyzoo/zoo/examples/objectdetection/predict.py
@@ -18,7 +18,6 @@ import argparse
 import cv2
 
 from zoo.common.nncontext import get_nncontext
-from zoo.feature.image.imageset import *
 from zoo.models.image.objectdetection.object_detector import *
 
 sc = get_nncontext(create_spark_conf().setAppName("Object Detection Example"))
