+++ pyzoo/zoo/feature/common.py
@@ -15,6 +15,7 @@
 #
 
 from bigdl.util.common import *
+
 from zoo.feature.image.imageset import ImageSet
 
 if sys.version >= '3':
@@ -86,22 +87,6 @@ class MLlibVectorToTensor(Preprocessing):
     """
     def __init__(self, size, bigdl_type="float"):
         super(MLlibVectorToTensor, self).__init__(bigdl_type, size)
-
-
-class ImageFeatureToTensor(Preprocessing):
-    """
-    a Transformer that convert ImageFeature to a Tensor.
-    """
-    def __init__(self, bigdl_type="float"):
-        super(ImageFeatureToTensor, self).__init__(bigdl_type)
-
-
-class RowToImageFeature(Preprocessing):
-    """
-    a Transformer that converts a Spark Row to a BigDL ImageFeature.
-    """
-    def __init__(self, bigdl_type="float"):
-        super(RowToImageFeature, self).__init__(bigdl_type)
 
 
 class FeatureLabelPreprocessing(Preprocessing):
