+++ zoo/src/main/scala/com/intel/analytics/zoo/examples/nnframes/imageInference/ImageInferenceExample.scala
@@ -20,7 +20,6 @@ import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric.NumericF
 import com.intel.analytics.bigdl.utils.LoggerFilter
 import com.intel.analytics.zoo.pipeline.nnframes._
 import com.intel.analytics.zoo.common.NNContext
-import com.intel.analytics.zoo.feature.common.{ImageFeatureToTensor, RowToImageFeature}
 import com.intel.analytics.zoo.feature.image._
 import org.apache.spark.sql.Row
 import org.apache.spark.sql.functions._
