+++ zoo/src/main/scala/com/intel/analytics/zoo/feature/image/ImageBrightness.scala
@@ -16,7 +16,6 @@
 package com.intel.analytics.zoo.feature.image
 
 import com.intel.analytics.bigdl.transform.vision.image.ImageFeature
-import com.intel.analytics.zoo.feature.common.{ImageProcessing}
 import com.intel.analytics.bigdl.transform.vision.image.augmentation
 
 /**
