+++ zoo/src/main/scala/com/intel/analytics/zoo/feature/image/ImageCenterCrop.scala
@@ -16,7 +16,6 @@
 package com.intel.analytics.zoo.feature.image
 
 import com.intel.analytics.bigdl.transform.vision.image.{ImageFeature, augmentation}
-import com.intel.analytics.zoo.feature.common.{ImageProcessing}
 class ImageCenterCrop(
     cropWidth: Int,
     cropHeight: Int,
