+++ zoo/src/main/scala/com/intel/analytics/zoo/feature/image/ImageHue.scala
@@ -15,7 +15,6 @@
  */
 package com.intel.analytics.zoo.feature.image
 
-import com.intel.analytics.zoo.feature.common.{ImageProcessing}
 import com.intel.analytics.bigdl.transform.vision.image.augmentation
 import com.intel.analytics.bigdl.transform.vision.image.ImageFeature
