+++ zoo/src/main/scala/com/intel/analytics/zoo/feature/image/ImageSetToSample.scala
@@ -17,7 +17,6 @@ package com.intel.analytics.zoo.feature.image
 
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.transform.vision.image.{ImageFeature, ImageFrameToSample}
-import com.intel.analytics.zoo.feature.common.ImageProcessing
 
 import scala.reflect.ClassTag
