+++ zoo/src/main/scala/com/intel/analytics/zoo/feature/image/RowToImageFeature.scala
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.analytics.zoo.feature.common
+package com.intel.analytics.zoo.feature.image
 
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.transform.vision.image.ImageFeature
+import com.intel.analytics.zoo.feature.common.Preprocessing
 import com.intel.analytics.zoo.pipeline.nnframes.NNImageSchema
 import org.apache.spark.sql.Row
