+++ zoo/src/main/scala/com/intel/analytics/zoo/pipeline/nnframes/python/PythonNNFrames.scala
@@ -26,6 +26,7 @@ import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.visualization.{TrainSummary, ValidationSummary}
 import com.intel.analytics.bigdl.{Criterion, Module}
 import com.intel.analytics.zoo.feature.common._
+import com.intel.analytics.zoo.feature.image.{ImageFeatureToTensor, RowToImageFeature}
 import com.intel.analytics.zoo.pipeline.nnframes._
 import org.apache.spark.api.java.JavaSparkContext
 import org.apache.spark.sql.DataFrame
